\name{Feldt1989}
\alias{Feldt1989}
\docType{data}
\title{Feldt's Numerical Example With 4 Items}
\description{
This covariance matrix was used as a numerical example in Feldt and Brennans' chapter in Educational Measurement titled \emph{Reliability}.
}
\usage{data(Feldt1989)}
\format{
  A covariance matrix of 4 items.
  }
\keyword{datasets}
