\name{oneFcong}
\alias{oneFcong}
\docType{data}
\title{One Factor Congeneric Covariance Matrix}
\description{
This Covariance matrix was used as the population model for one set of simulations.  It was used to represent a congeneric data structure in which the factor loadings are set at .5, .6, .7, .8, .5, .6, .7, and .8.  The error variances were set at .6^2, .7^2, .8^2, .9^2, .6^2, .7^2, .8^2, and .9^2.
}
\usage{data(oneFcong)}
\format{
  A covariance matrix of 8 theoretical items.
}
\examples{
data(oneFcong)
}
\keyword{datasets}
