\name{oneFpar}
\alias{oneFpar}
\docType{data}
\title{One Factor Parallel Covariance Matrix}
\description{
This Covariance matrix was used as the population model for one set of simulations.  It was used to represent a parallel data structure in which all factor loadings and error variances are set at .6.
}
\usage{data(oneFpar)}
\format{
  A covariance matrix of 8 theoretical items.
}
\examples{
data(oneFpar)
}
\keyword{datasets}
