\name{oneFtau}
\alias{oneFtau}
\docType{data}
\title{One Factor Tau-Equivalent Covariance Matrix}
\description{
This covariance matrix was used as the population model for one set of simulations.  It was used to represent a tau equivalent data structure in which the factor loadings are set at .6.  The error variances were set at .6^2, .7^2, .8^2, .9^2, .6^2, .7^2, .8^2, and .9^2.
}
\usage{data(oneFtau)}
\format{
  A covariance matrix of 8 theoretical items.
}
\examples{
data(oneFtau)
}
\keyword{datasets}
