\name{Rosenberg}
\alias{Rosenberg}
\docType{data}
\title{Rosenberg Self-Esteem}
\description{
The data set was collected in Southern Utah in the Fall of 2010.  The investigation sought responses from high school and college students.  It should be noted that the reverse coded items have been flipped.
}
\usage{data(Rosenberg)}
\format{
  A data frame with 837 observations on the following 10 variables.
  \describe{
    \item{\code{SEFailureR}}{All in all, I am inclined to feel that I am a failure.}
    \item{\code{SENoGoodR}}{At times I think I am no good at all.}
    \item{\code{SEAble}}{I am able to do things as well as most other people.}
    \item{\code{SEUselessR}}{I certainly feel useless at times.}
    \item{\code{SENoProudR}}{I feel I do not have much to be proud of.}
    \item{\code{SEGoodQualities}}{I feel that I have a number of good qualities.}
    \item{\code{SEWorth}}{I feel that I am a person of worth, at least on an equal plane with others.}
    \item{\code{SEPositive}}{I take a positive attitude toward myself.}
    \item{\code{SERespectR}}{I wish I could have more respect for myself.}
    \item{\code{SESatisfied}}{On the whole, I am satisfied with myself.}
  }
}
\examples{
data(Rosenberg)
}
\keyword{datasets}
