\name{TenBerge2004
}
\alias{TenBerge2004}
\docType{data}
\title{
De Leeuw (1983) Political Survey Items
}
\description{Six political survey items, N = 119, and unidensional.}
\usage{data(TenBerge2004)}
\details{This is a covariance matrix that comes De Leeuw (1983) and represents six political survey items.  These items are based on N = 119 members of parliament and are supposed to measure the same trait and be unidensional.}
\source{
Ten Berge, J. M., & Socan, G. (2004). The greatest lower bound to the reliability of a test and the hypothesis of unidimensionality. Psychometrika, 69(4), 613-625.
}
\references{
De Leeuw, J. (1983). Models and methods for the analysis of correlation coefficients. Journal of Econometrics, 22, 113-137.
}
\examples{
data(TenBerge2004)
}
\keyword{datasets}
