\name{WAIS1955}
\alias{WAIS1955}
\docType{data}
\title{
Wechsler Adult Intellegence Scale (1955)
}
\description{Data comes from Warner, Meeker, and Eels (1960) and is a multidimensional scale of composite scores of social class.}
\usage{data(Warner)}
\format{
  The format is:
 num [1:6, 1:6] 1 0.87 0.76 0.71 0.7 0.77 0.87 1 0.82 0.81 ...
}
\details{ This is a covariance matrix of the 11 subtests of the WAIS based on N = 300.  The subtests include: comprehension, arithmetic, similarities, digit span, vocabulary, digit symbol, picture completion, block design, picture arrangement, and object assembly.  These data have been used by Bentler (1972) to show the stark difference between alpha and the glb.}
\source{
Bentler, PM (1972). "A Lower-Bound Method for the Dimension-Free Measurement of Internal Consistency" 1(4), 343-357.
}
\references{
Wechsler, D. (1955), "Manual for the Wechsler Adult Intelligence Scale," The Psychological Corporation, New York
}
\examples{
data(WAIS1955)
}
\keyword{datasets}
