\name{Guttman}
\alias{Guttman}
\docType{data}
\title{Artificial Guttman Scale}
\description{This artificial Guttman scale was taken from Bentler (1972)
}
\usage{data(Guttman)}
\format{
  The format is:
 num [1:5, 1:5] 0.09 0.08 0.05 0.02 0.01 0.08 0.16 0.1 0.04 0.02 ...
}
\source{
Bentler, PM (1972). "A Lower-Bound Method for the Dimension-Free Measurement of Internal Consistency" 1(4), 343-357.
}
\examples{
data(Guttman)
}
\keyword{datasets}
